refactor(test): one loop for waiting on a contended backend, not two
Two suites wait for a racer that is provably contended, and they ask genuinely
different questions: approvals polls pg_stat_activity for a backend blocked by
a known pid, the stage-removal race polls FOR UPDATE NOWAIT until 55P03. Those
questions are not redundant and neither is deleted.

What was duplicated is everything around them — the budget, the pacing, the
select over racer-finished / budget-expired / look-again, and the rule that a
probe which gave up must say what the run failed to prove rather than pass
having proved nothing.

They had already drifted, which is the argument made concrete: one budget was
90 seconds and the other 60, under a comment citing the first as the reasoning
for the second. Neither number was wrong. The point is that only one of them
could be corrected.

testdb.WaitForContention takes the look — func(context.Context) (bool, error),
which was already approvals' shape — and each caller supplies its own question
and its own two failure sentences. The reasoning for the budget and the pacing
lives with the loop, where correcting it corrects both.

The suites' own tests carry the property, unchanged: the approvals package's
proof that its probe sees a backend dialled after its first look, and the
stage-removal race's two orderings.
